# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: COMPAON API
  version: 1.0.0
  contact:
    email: compaon@dalavina.com
    name: Compaon
    url: 'https://dalavina.com/compaon'
  license:
    name: EULA
    url: 'https://en.wikipedia.org/wiki/End-user_license_agreement'
  description: API definition for COMPAON
servers:
  - url: 'http://52.74.156.150:3000/api/v1'
paths:
...
  '/space-templates/{id}':
    get:
      security:
      - bearerAuth: []
      summary: Get space template by ID
      description: Return a single space template
      operationId: getSpaceTemplateById
      parameters:
        - name: id
          description: ID to get space template
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'space_template_positions'
          description: |
            - space_template_positions
            - space_template_locations
            - spaces
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceTemplate'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Space Template
    delete:
      tags:
        - Space Template
      security:
      - bearerAuth: []
      summary: Delete an existing space template
      description: Delete an existing space template by Id
      operationId: deleteSpaceTemplate
      parameters:
        - name: id
          description: ID to delete space template
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/space-templates':
    get:
      security:
      - bearerAuth: []
      summary: Show list space templates
      description: Show list space templates
      operationId: getSpaceTemplates
      parameters:
      - in: query
        name: page
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '1'
        description: The number of items to skip before starting to collect the result set
      - in: query
        name: limit
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '10'
        description: The numbers of items to return
      - in: query
        name: search
        schema:
          type: string
          example: 'Space 01'
        description:
         " Searching on allow columns:['id', 'name']\n
          example: http://localhost:3000/api/v1/space-templates?search=1"
      - in: query
        name: sort_by
        allowReserved: true
        schema:
          type: string
          example: 'name:DESC'
        description: "Sort on allow columns:[['id', 'name', 'created_at'],[DESC,ASC]]\n default column:['created_at', 'DESC']"
      - in: query
        name: filter
        allowReserved: true
        schema:
          type: array
          items:
            type: string
            example: 'name:$eq:Space%2001'
        description: Filter by the field with syntax `column_name:operator:value`
      - in: query
        name: populate
        schema:
          type: array
          items:
            type: string
            example: 'space_template_positions'
        description: |
          - space_template_positions
          - spaces
      - in: query
        name: withDeleted
        schema:
          type: string
          example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSpaceTemplates'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Space Template
    post:
      security:
      - bearerAuth: []
      tags:
        - Space Template
      summary: Add a space template
      description: Add a space template
      operationId: addSpaceTemplate
      requestBody:
        description: Create a space template
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceTemplate'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceTemplate'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/spaces/{id}':
    get:
      security:
       - bearerAuth: []
      summary: Get space by ID
      description: Return a single space
      operationId: getSpaceById
      parameters:
      - name: id
        description: ID to update space
        in: path
        required: true
        schema:
          type: string
          pattern: '^[0-9]+$'
      - in: query
        name: populate
        schema:
          type: array
          items:
            type: string
            example: 'space_template'
        description: |
            List of populatable columns:
            - exhibition
            - space_template
            - space_template.space_template_positions
            - space_images
            - space_images.image
            - space_images.space_template_position
            - space_videos
            - space_videos.video
            - space_videos.space_template_position
      - in: query
        name: withDeleted
        schema:
          type: string
          example: 'true'
      tags:
        - Space
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    put:
      security:
      - bearerAuth: []
      tags:
        - Space
      summary: Update an existing space
      description: Update an existing space by Id
      operationId: updateSpace
      parameters:
        - name: id
          description: ID to update space
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Update an existing space
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Space'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not found
                    details: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
...
components:
  schemas:
...
    Space:
      type: object
      required:
        - id
        - name
        - exhibition_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        space_template_id:
          type: integer
          example: 1
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: 'Mau khong gian tien sanh mac dinh'
        exhibition_id:
          type: number
          example: 1
        locations:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/Location'
        space_images:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/SpaceImage'
        space_videos:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/SpaceVideo'
    SpaceVideo:
      type: object
      required:
        - id
        - space_id
        - space_template_position_id
        - video_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        space_id:
          type: number
          writeOnly: true
          example: 1
        space_template_position_id:
          type: number
          writeOnly: true
          example: 1
        select_media_id:
          type: number
          writeOnly: true
          example: 1
        video_id:
          type: number
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: "base64 string"
    SpaceImage:
      type: object
      required:
        - id
        - space_id
        - space_template_position_id
        - image_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        space_id:
          type: number
          writeOnly: true
          example: 1
        space_template_position_id:
          type: number
          writeOnly: true
          example: 1
        select_media_id:
          type: number
          writeOnly: true
          example: 1
        image_id:
          type: number
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: "base64 string"
    Location:
      type: object
      required:
        - id
        - status
        - space_id
        - space_template_location_id
      properties:
        id:
          type: integer
          readOnly: true
        status:
          type: string
          enum: ['available', 'reserved']
          default: 'available'
        space_id:
          type: number
          writeOnly: true
          example: 1
        space_template_location_id:
          type: number
          writeOnly: true
          example: 1
        space_template_location:
          $ref: '#/components/schemas/SpaceTemplateLocation'
    SpaceTemplateLocation:
      type: object
      required:
        - id
        - name
        - position_x
        - position_y
        - position_z
        - rotation_x
        - rotation_y
        - rotation_z
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          example: 'a'
        position_x:
          type: number
          example: 1
        position_y:
          type: number
          example: 1
        position_z:
          type: number
          example: 1
        rotation_x:
          type: number
          example: 1
        rotation_y:
          type: number
          example: 1
        rotation_z:
          type: number
          example: 1
...
    SpaceTemplate:
      type: object
      required:
        - id
        - name
        - model_data
        - thumbnail_data
        - thumbnail_id
        - created_by
        - created_date
        - model_id
        - map_id
        - map_data
        - space_template_positions
        - space_template_locations
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Mẫu không gian 01"
        created_by:
          type: number
          example: 1
          readOnly: true
        model_id:
          type: integer
          readOnly: true
        model_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        thumbnail_id:
          type: integer
          readOnly: true
        thumbnail_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        map_id:
          type: integer
          readOnly: true
        map_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        created_date:
          type: string
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-01-01T00:00:00.000Z'
        space_template_positions:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/SpaceTemplatePosition'
        space_template_locations:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/SpaceTemplateLocation'
...
    SpaceTemplatePosition:
      type: object
      required:
       - id
       - position
       - type
      properties:
        id:
          type: integer
          readOnly: true
        position:
          type: string
          maxLength: 255
        type:
          type: string
          enum: ['image','video']
    PaginatedSpaceTemplates:
      type: object
      required:
        - page
        - limit
        - total
        - space_templates
      properties:
        page:
          type: integer
          example: 0
          description: The number of items to skip before starting to collect the result set.
        limit:
          type: integer
          example: 10
          description: The number of items to return.
        total:
          type: integer
          format: int64
          example: 1
        space_templates:
          type: array
          items:
            $ref: '#/components/schemas/SpaceTemplate'
...
    ErrorResponseBody:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT